When you remove user/deck remove links
In the database we link users to decksets, and deck to cards. If we
remove these IDs from the database but not the links, if the ID gets
reused they will pick up the old users cards. Make sure to delete the
links when we delete the items

Simplify this by having Deck has a remove function